feat: 安全加固 + 文件传输全链路 + 协议统一 + 测试恢复
P0 安全:
- 管理面认证: ControlServer/Web 管理端令牌认证, admin_login, 改密轮换令牌,
  CORS/Origin/Host 校验, 请求体上限
- 更新器: Ed25519 清单签名, zip 路径穿越/炸弹防护, apply_update.js 消除 cmd 注入,
  强制 sha256+size 校验, 停止清空 localStorage
- 语音: XChaCha20 nonce 跨实例唯一(随机16B前缀), 中继频道成员校验+身份绑定,
  删除原样转发(fail-closed), shared_ptr 消除 UAF, 视频密钥日志移除,
  内存池释放竞争修复, TCP 载荷读取超时
- 登录: 失败限速, max_users 检查修复, 密钥轮换回写
- 三端删除明文密钥回退(失败即断开), Android 伪加密 fallback 清理, cleartext 关闭

协议统一:
- 自定义小端 TLV 确认为唯一运行时线格式(三端逐字段一致)
- 金样互操作测试双向锁定(C++<->Python), docs/protocol-wire-format.md
- Android MessageType 补齐 44/45/60-62/70-71/80-83
- proto 追加文件分片消息 44-47(只追加不改编号)

文件传输:
- 服务端: 分片接收/写盘/ACK, 分片下载, 归属与路径穿越校验
- 客户端: 真实字节流上传/取回/重组/缓存, 聊天中继分片通道
- e2e 冒烟: 上传->落盘->下载->重组字节一致

测试: ctest 450->459 全绿(中继安全5/互操作金样6), Python 30 用例
质量: logging_setup 共享模块, 密钥材料日志移除, Python 解析器健壮性,
      voice_crypto nonce 对称修复, 平台宏 ws2_32/_USE_MATH_DEFINES 修复
清理: 解除跟踪 PyInstaller 产物, 忽略 .mimosa/.tmp/.nevo_update

+++ mobile/android/app/src/main/java/com/nevo/voip/core/database/entity/ChatMessageEntity.kt
@@ -14,7 +14,7 @@ import androidx.room.PrimaryKey
 data class ChatMessageEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
-    val channelId: Int,
+    val channelId: Long,
     val senderId: Long,
     val senderName: String,
     val content: String,
